+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 8.1
 - Fix SQL error when Student / Staff ID is hacked / not an integer in URL in Current.php
 - SQL ACCOUNTING_SALARIES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 - Add File Attached to Salaries in Salaries.php & Accounting/functions.inc.php
+- SQL BILLING_FEES table: Add FILE_ATTACHED column in Update.fnc.php & rosariosis.sql
 
 Changes in 8.0.4
 ----------------